@@ -5,12 +5,12 @@ description: >-
 publish_to: none
 
 environment:
-  sdk: ">=2.5.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
-  intl: ^0.16.0
-  path: ^1.6.1
+  intl: ^0.17.0
+  path: ^1.8.2
 
 dev_dependencies:
-  build_runner: ^2.0.0
-  build_web_compilers: ^3.0.0
+  build_runner: ^2.1.11
+  build_web_compilers: ^3.2.3
